image_upscaling_ai/requirements.txt
# Image Upscaling AI Requirements

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pillow>=10.0.0
numpy>=1.24.0

# HTTP client
httpx>=0.25.0

# Image processing
opencv-python-headless>=4.8.0

# Real-ESRGAN (optional, for best quality upscaling)
# Install with: pip install realesrgan basicsr
# realesrgan>=0.3.0
# basicsr>=1.4.2

# Optional: For advanced image processing
# scikit-image>=0.21.0
# imageio>=2.31.0

# Performance monitoring
psutil>=5.9.0

